requirements.txt
jupyter
pandas
openpyxl
pyxlsb
pytest
python-dotenv
numpy
pyspark
supabase